feat: save order fee with each order

+++ db/models/models.go
@@ -51,13 +51,13 @@ type Store struct {
 	Description         string         `db:"description"`
 	Address             string         `db:"address"`
 	AveragePickupTime   int            `db:"average_pickup_time"`
-	AverageReview       float64        `db:"average_review"`
+	AverageReview       float32        `db:"average_review"`
 	ReviewCount         int            `db:"review_count"`
 	GoogleMapsLink      string         `db:"google_maps_link"`
 	PhoneNumber         string         `db:"phone_number"`
 	StripeAccountId     sql.NullString `db:"stripe_account_id"`
 	StripeAccountStatus sql.NullString `db:"stripe_account_status"`
-	Fee                 float64        `db:"fee"`
+	Fee                 float32        `db:"fee"`
 	IsOpen              bool           `db:"is_open"`
 	ImageUrl            string         `db:"image_url"`
 	MerchantId          int            `db:"merchant_id"`
